src/requirements.txt
azure-cosmos
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
pytest